requirements.txt
asgiref==3.11.0
Django==5.2.9
sqlparse==0.5.4
tzdata==2025.3
psycopg2-binary==2.9.10
djangorestframework==3.15.2
drf-spectacular==0.27.2
python-dotenv==1.0.1
pyotp==2.9.0
